Clean up handling of clock_flag and the ifclock macro.
This commit makes certain that the block following the
ifclock is executed exactly once if the flag bit is set,
and makes certain that the bit is not tested a second time.

Also, this commit makes certain that clearing the clock_flag bit
does not clear any other bit by mistake (by disabling interrupts
around the clearing of the bit).